Split up ParameterDefinition
